HomeAway - Selenium Automation
This is the project to automate a website using Selenium.
Consists of 3 tests as per specification:
- Submit an order for an Apple iPhone4s 16GB SIM-Free – Black (known
issue with State drop-down, selecting Country is adequate) and verify
the Total Price: given is correct (assume shipping cost is correct based
on your choice). You may assume prices shown on product pages are the
correct price.
- Verify updating your account details is saved and retrieved after
logging out and back in using the My Account link.
- Verify removing all items from your cart produces an empty cart
message.